meson: complete sourcemap support

+++ scripts/gh-pages-deploy.sh
@@ -6,8 +6,6 @@ if [ ! -e "src/api/api.h" ]; then
   echo "Please run this script from the root directory of Lite XL."; exit 1
 fi
 
-source scripts/common.sh
-
 # find file packager to package the data dirs
 if ! command -v file_packager >/dev/null 2>&1; then
     if ! command -v emcc >/dev/null 2>&1; then
@@ -24,21 +22,28 @@ if ! command -v file_packager >/dev/null 2>&1; then
     export PATH="$(readlink -f "$(dirname "${emcc_abs_path}")/tools"):$PATH"
 fi
 
-build_dir=$(get_default_build_dir "emscripten" "wasm32")
+build_dir="build-wasm32-emscripten"
 dest_dir=github-pages
 
 if [[ ! -d "$build_dir" ]]; then
-    if [[ -n "$CI" ]]; then
-        DISABLE_DWARF="-Dwasm_dwarf=false"
+    args=(
+        '-Dwasm_preload_files=false'
+        '-Dwasm_build_bundle=true'
+    )
+    if [[ "$PRODUCTION" = "true" ]]; then
+        args+=('-Dwasm_debug=sourcemap')
+    else
+        args+=('-Dwasm_debug=dwarf')
     fi
-    meson setup "$build_dir" --cross-file resources/cross/unknown-wasm32.txt -Dwasm_preload_files=false -Dwasm_build_bundle=true $DISABLE_DWARF
+    meson setup "$build_dir" --cross-file resources/cross/unknown-wasm32.txt "${args[@]}"
 fi
 
-meson compile -C "$build_dir"
+ninja -C "$build_dir" $1
 
-if [ -d "${dest_dir}" ]; then
-    find "${dest_dir}" -mindepth 1 -delete
-fi
-DESTDIR="$(pwd)/${dest_dir}" meson install --no-rebuild --skip-subprojects --tags shell,bundle,exe -C "${build_dir}"
-
+if [[ -z "$1" ]]; then
+    if [ -d "${dest_dir}" ]; then
+        find "${dest_dir}" -mindepth 1 -delete
+    fi
+    DESTDIR="$(pwd)/${dest_dir}" meson install --no-rebuild --skip-subprojects --tags shell,bundle,exe,sourcemap -C "${build_dir}"
 rm -f "${dest_dir}/lite-xl.js"
+fi
